requirements.txt
flask==2.3.3
flask-cors==4.0.0
selenium==4.15.0
deep-translator==1.11.4
pypdf==3.17.0
python-pptx==0.6.23
langchain==0.0.350
transformers==4.36.2
torch==2.1.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.3.9
tiktoken==0.5.2
chromedriver-autoinstaller==0.6.2
gunicorn==21.2.0